requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
python-dotenv==1.0.1
markdown2==2.5.1
Jinja2==3.1.5
GitPython==3.1.0
pandas==2.2.3
markitdown==0.0.1a3
gitdb==4.0.11
python-gitlab==5.2.0
Whoosh==2.7.4
